--- a/be/src/formats/orc/apache-orc/c++/src/Reader.cc
+++ b/be/src/formats/orc/apache-orc/c++/src/Reader.cc
@@ -428,14 +428,9 @@ void RowReaderImpl::loadStripeIndex() {
 
     // obtain row indexes for selected columns
     uint64_t offset = currentStripeInfo.offset();
-
-    // usually row index size is small.
-    uint64_t rowIndexSize = currentStripeInfo.indexlength();
-    contents->stream->prepareCache(InputStream::PrepareCacheScope::READ_FULL_ROW_INDEX, offset, rowIndexSize);
     for (int i = 0; i < currentStripeFooter.streams_size(); ++i) {
         const proto::Stream& pbStream = currentStripeFooter.streams(i);
         uint64_t colId = pbStream.column();
-        // We only need to load active column's RowIndex
         if (selectedColumns[colId] && pbStream.has_kind() &&
             (pbStream.kind() == proto::Stream_Kind_ROW_INDEX ||
              pbStream.kind() == proto::Stream_Kind_BLOOM_FILTER_UTF8)) {
@@ -1020,7 +1015,16 @@ void RowReaderImpl::buildIORanges(std::vector<InputStream::IORange>* io_ranges) 
         // ColumnId = 0 is root column, we always need it
         if (columnId == 0 || selectedColumns[columnId] || lazyLoadColumns[columnId]) {
             bool is_active = true;
-            if (lazyLoadColumns[columnId]) {
+
+            // We didn't support stripe index lazy load, so we will regard all index stream as active column
+            bool is_stripe_index = false;
+            if (stream.has_kind() && (stream.kind() == proto::Stream_Kind_ROW_INDEX ||
+                                      stream.kind() == proto::Stream_Kind_BLOOM_FILTER_UTF8)) {
+                is_stripe_index = true;
+            }
+
+            // we only seperate io range for column's data, don't include column's index
+            if (!is_stripe_index && lazyLoadColumns[columnId]) {
                 is_active = false;
             }
             io_ranges->emplace_back(InputStream::IORange{.offset = offset, .size = length, .is_active = is_active});
@@ -1065,17 +1069,18 @@ void RowReaderImpl::startNextStripe() {
             }
         }
 
-        contents->stream->prepareCache(InputStream::PrepareCacheScope::READ_FULL_STRIPE, currentStripeInfo.offset(),
-                                       stripeSize);
+        // release previous stripe's io ranges
         if (isIOCoalesceEnabled) {
-            contents->stream->clearIORanges();
+            contents->stream->releaseToOffset(currentStripeInfo.offset());
         }
         currentStripeFooter = getStripeFooter(currentStripeInfo, *contents);
         rowsInCurrentStripe = currentStripeInfo.numberofrows();
-        if (isIOCoalesceEnabled) {
+        // We need to check this stripe is already set in shared buffer(tiny stripe optimize) to avoid shared buffer overlap
+        if (isIOCoalesceEnabled &&
+            !contents->stream->isAlreadyCollectedInSharedBuffer(currentStripeInfo.offset(), stripeSize)) {
             std::vector<InputStream::IORange> io_ranges;
             buildIORanges(&io_ranges);
-            contents->stream->setIORanges(io_ranges, true);
+            contents->stream->setIORanges(io_ranges);
         }
 
         if (sargsApplier) {
@@ -1146,6 +1151,11 @@ void RowReaderImpl::startNextStripe() {
             // advance to next stripe when current stripe has no matching rows
             currentStripe += 1;
             currentRowInStripe = 0;
+
+            if (contents->stream->get_lazy_column_coalesce_counter() != nullptr) {
+                // Skip entrie stripe, which means we didn't need to coalesce active and lazy column together
+                contents->stream->get_lazy_column_coalesce_counter()->fetch_sub(1, std::memory_order_relaxed);
+            }
         } else {
             break;
         }
@@ -1448,7 +1458,6 @@ std::unique_ptr<Reader> createReader(std::unique_ptr<InputStream> stream, const 
             throw ParseError("File size too small");
         }
         std::unique_ptr<DataBuffer<char>> buffer(new DataBuffer<char>(*contents->pool, readSize));
-        stream->prepareCache(InputStream::PrepareCacheScope::READ_FULL_FILE, 0, fileLength);
         stream->read(buffer->data(), readSize, fileLength - readSize);
 
         postscriptLength = buffer->data()[readSize - 1] & 0xff;
@@ -1555,8 +1564,6 @@ uint64_t InputStream::getNaturalReadSizeAfterSeek() const {
     return 128 * 1024;
 }
 
-void InputStream::prepareCache(PrepareCacheScope scope, uint64_t offset, uint64_t length) {}
-
 bool InputStream::isIOCoalesceEnabled() const {
     return false;
 }
@@ -1565,8 +1572,16 @@ bool InputStream::isIOAdaptiveCoalesceEnabled() const {
     return false;
 }
 
-void InputStream::clearIORanges() {}
+bool InputStream::isAlreadyCollectedInSharedBuffer(const int64_t offset, const int64_t length) const {
+    return false;
+}
 
-void InputStream::setIORanges(std::vector<InputStream::IORange>& io_ranges, const bool is_from_stripe) {}
+void InputStream::releaseToOffset(const int64_t offset) {}
+
+void InputStream::setIORanges(std::vector<InputStream::IORange>& io_ranges) {}
+
+std::atomic<int32_t>* InputStream::get_lazy_column_coalesce_counter() {
+    return nullptr;
+}
 
 } // namespace orc