--- a/be/src/formats/orc/utils.h
+++ b/be/src/formats/orc/utils.h
@@ -14,10 +14,7 @@
 
 #pragma once
 
-#include <exception>
 #include <orc/OrcFile.hh>
-#include <set>
-#include <unordered_map>
 #include <utility>
 
 #include "cctz/civil_time.h"
@@ -26,12 +23,58 @@
 #include "formats/orc/orc_mapping.h"
 #include "formats/orc/utils.h"
 #include "gen_cpp/orc_proto.pb.h"
-#include "runtime/types.h"
+#include "io/shared_buffered_input_stream.h"
 #include "types/date_value.h"
-#include "types/logical_type.h"
 #include "types/timestamp_value.h"
 
 namespace starrocks {
+
+class DiskRange {
+public:
+    DiskRange(int64_t off, int64_t len) : offset(off), length(len) {
+        DCHECK(off >= 0);
+        DCHECK(len > 0);
+    }
+
+    /**
+    * Returns the minimal DiskRange that encloses both this DiskRange
+    * and otherDiskRange. If there was a gap between the ranges the
+    * new range will cover that gap.
+    */
+    DiskRange span(const DiskRange& otherDiskRange) const {
+        const int64_t start = std::min(offset, otherDiskRange.offset);
+        const int64_t end = std::max(get_end(), otherDiskRange.get_end());
+        return DiskRange(start, end - start);
+    }
+
+    int64_t get_end() const { return offset + length; }
+
+    int64_t offset;
+    int64_t length;
+};
+
+class DiskRangeHelper {
+public:
+    static void mergeAdjacentDiskRanges(std::vector<io::SharedBufferedInputStream::IORange>& io_ranges,
+                                        const std::vector<DiskRange>& disk_ranges, const int64_t max_merge_distance,
+                                        const int64_t max_merged_size) {
+        if (disk_ranges.empty()) {
+            return;
+        }
+        DiskRange last = disk_ranges[0];
+        for (size_t i = 1; i < disk_ranges.size(); i++) {
+            DiskRange current = disk_ranges[i];
+            DiskRange merged = last.span(current);
+            if (merged.length <= max_merged_size && last.get_end() + max_merge_distance >= current.offset) {
+                last = merged;
+            } else {
+                io_ranges.emplace_back(last.offset, last.length, true);
+                last = current;
+            }
+        }
+        io_ranges.emplace_back(last.offset, last.length, true);
+    }
+};
 
 // Hive ORC char type will pad trailing spaces.
 // https://docs.cloudera.com/documentation/enterprise/6/6.3/topics/impala_char.html