## Why I'm doing:
The original implementation was not well integrated with the shared buffer(io coalesce), and the optimization in some places was too radical, which would lead read more data.

## What I'm doing:
* Remove prepare cache for full file, it may occur bad case when a predicate has strong selectivity. Sometime these case only need to read ORC's footer, but we will load the whole file.
* Remove prepare cache for row index, because we already collected row index's io range, shared buffer can handle it well.
* Move tiny stripe optimization into shared buffer, so SR's io coalesce policy can handle it well.
* Move tiny stripe optimization codes out of apache orc.

**Previous profile:**
```bash
 - InputStream: 
   - AppIOBytesRead: 1.530 GB
   - AppIOCounter: 322
   - AppIOTime: 5s900ms
   - FSIOBytesRead: 3.014 GB
   - FSIOCounter: 546
   - FSIOTime: 5s225ms
- SharedBuffered: 
 - DirectIOBytes: 9.042 MB
 - DirectIOCount: 62
 - DirectIOTime: 1s865ms
 - SharedIOBytes: 3.006 GB
 - SharedIOCount: 484
 - SharedIOTime: 3s587ms
- ORC: 
 - IcebergV2FormatTimer: 
   - DeleteFileBuildFilterTime: 0ns
   - DeleteFileBuildTime: 0ns
   - DeleteFilesPerScan: 0
 - StripeActiveLazyColumnIOCoalesceSeperately: 0
 - StripeActiveLazyColumnIOCoalesceTogether: 256
 - TotalStripeNumber: 256
 - TotalStripeSize: 1.548 GB
```

**Now profile:**
```bash
 - InputStream: 
   - AppIOBytesRead: 8.533 MB
   - AppIOCounter: 52.284K (52284)
   - AppIOTime: 2s55ms
   - FSIOBytesRead: 1.621 GB
   - FSIOCounter: 348
   - FSIOTime: 1s854ms
- SharedBuffered: 
 - DirectIOBytes: 14.502 MB
 - DirectIOCount: 90
 - DirectIOTime: 149.129ms
 - SharedIOBytes: 1.607 GB
 - SharedIOCount: 258
 - SharedIOTime: 1s822ms
- ORC: 
 - IcebergV2FormatTimer: 
   - DeleteFileBuildFilterTime: 0ns
   - DeleteFileBuildTime: 0ns
   - DeleteFilesPerScan: 0
 - StripeActiveLazyColumnIOCoalesceSeperately: 0
 - StripeActiveLazyColumnIOCoalesceTogether: 30
 - TotalStripeNumber: 256
 - TotalStripeSize: 1.548 GB
 - TotalTinyStripeSize: 1.360 GB
```

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [x] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5